target_name -> target_object. Added ggt_props for Actions.
